# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Otacon Phone Automation API",
        "description": "REST API for controlling Android phones connected to a Raspberry Pi.",
        "license": {
            "name": "MIT"
        },
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "https://otacon-pi.<tailnet>.ts.net:8080",
            "description": "Default Tailscale address"
        }
    ],
    "paths": {
...
        "/api/apns": {
            "get": {
                "tags": [
                    "APN"
                ],
                "operationId": "listApns",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ApnOverride"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "APN"
                ],
                "operationId": "createApn",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApnBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApnOverride"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/apns/enabled": {
            "get": {
                "tags": [
                    "APN"
                ],
                "operationId": "getApnsEnabled",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApnEnabled"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "APN"
                ],
                "operationId": "setApnsEnabled",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SetApnEnabledBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApnEnabled"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/apns/{apn_id}": {
            "put": {
                "tags": [
                    "APN"
                ],
                "operationId": "updateApn",
                "parameters": [
                    {
                        "name": "apn_id",
                        "in": "path",
                        "description": "Android override APN id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ApnBody"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApnOverride"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "APN"
                ],
                "operationId": "deleteApn",
                "parameters": [
                    {
                        "name": "apn_id",
                        "in": "path",
                        "description": "Android override APN id",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int32"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ApnBody": {
                "type": "object",
                "properties": {
                    "apn": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "authType": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "entryName": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "mmsPort": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int32"
                    },
                    "mmsProxy": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "mmsc": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "operatorNumeric": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "password": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "protocol": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "roamingProtocol": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "types": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "user": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
            "ApnEnabled": {
                "type": "object",
                "required": [
                    "enabled"
                ],
                "properties": {
                    "enabled": {
                        "type": "boolean"
                    }
                }
            },
            "ApnOverride": {
                "type": "object",
                "required": [
                    "id",
                    "entryName",
                    "apnName",
                    "operatorNumeric",
                    "types",
                    "protocol",
                    "roamingProtocol",
                    "authType",
                    "enabled"
                ],
                "properties": {
                    "apnName": {
                        "type": "string"
                    },
                    "authType": {
                        "type": "string"
                    },
                    "enabled": {
                        "type": "boolean"
                    },
                    "entryName": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer",
                        "format": "int32"
                    },
                    "mmsPort": {
                        "type": [
                            "integer",
                            "null"
                        ],
                        "format": "int32"
                    },
                    "mmsProxy": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "mmsc": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "operatorNumeric": {
                        "type": "string"
                    },
                    "protocol": {
                        "type": "string"
                    },
                    "roamingProtocol": {
                        "type": "string"
                    },
                    "types": {
                        "type": "string"
                    },
                    "user": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                }
            },
...
            "SetApnEnabledBody": {
                "type": "object",
                "required": [
                    "enabled"
                ],
                "properties": {
                    "enabled": {
                        "type": "boolean"
                    }
                }
            },
...
        }
    }
}